Started creating a docker package (#34)
Moved the majority of the Docker Hub specific code to a 'docker'
package. I would estimate that 85% of the code was moved. The rest isn't
moved yet because I still want more time to really lock down the
structure/API of this new package. Since Sonar is still < v1.0.0, the
API promise doesn't apply yet.

I'd like to finalize and optimize the package in future PRs and move the
rest of the code before the v1 release.